+++ Editor/Generatable/Directives.cs
@@ -1,7 +1,0 @@
-namespace NPTP.UnitySourceGen.Editor.Generatable
-{
-    public static class Directives
-    {
-        public const string UNITY_ENGINE = "UnityEngine";
-    }
-}